+++ SwiftPMTests/objc-import-test/objc-header.m
@@ -21,9 +21,9 @@
 #import "FirebaseCore/FirebaseCore.h"
 #import "FirebaseCrashlytics/FirebaseCrashlytics.h"
 #import "FirebaseDatabase/FirebaseDatabase.h"
-#if defined(FirebaseFirestore) || defined(FIREBASE_FIRESTORE_ENABLED)
+#if defined(FirestoreSupport) || defined(FIREBASE_FIRESTORE_ENABLED)
 #import "FirebaseFirestore/FirebaseFirestore.h"
-#endif  // defined(FirebaseFirestore) || defined(FIREBASE_FIRESTORE_ENABLED)
+#endif  // defined(FirestoreSupport) || defined(FIREBASE_FIRESTORE_ENABLED)
 #import "FirebaseInstallations/FirebaseInstallations.h"
 #import "FirebaseMessaging/FirebaseMessaging.h"
 #if (TARGET_OS_IOS && !TARGET_OS_MACCATALYST) || TARGET_OS_TV
@@ -41,9 +41,9 @@
 #import <FirebaseCore/FirebaseCore.h>
 #import <FirebaseCrashlytics/FirebaseCrashlytics.h>
 #import <FirebaseDatabase/FirebaseDatabase.h>
-#if defined(FirebaseFirestore) || defined(FIREBASE_FIRESTORE_ENABLED)
+#if defined(FirestoreSupport) || defined(FIREBASE_FIRESTORE_ENABLED)
 #import <FirebaseFirestore/FirebaseFirestore.h>
-#endif  // defined(FirebaseFirestore) || defined(FIREBASE_FIRESTORE_ENABLED)
+#endif  // defined(FirestoreSupport) || defined(FIREBASE_FIRESTORE_ENABLED)
 #if TARGET_OS_IOS || TARGET_OS_TV
 #import <FirebaseInAppMessaging/FirebaseInAppMessaging.h>
 #endif